fuel_flask_server/src/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
Flask==3.0.0
greenlet==3.0.3
idna==3.6
ifaddr==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
psycopg2==2.9.9
py-eureka-client==0.11.8
pycparser==2.21
PyMySQL==1.1.0
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
